Add files via upload
Simple Todo List

This is a simple todo list application that is built using HTML, CSS, and JavaScript. The application allows users to create, edit, delete, and complete tasks.

Getting Started

To get started, simply open the index.html file in your browser. You will be presented with an empty todo list, where you can start adding tasks by clicking on the "Add Task" button.

Creating a Task

To create a new task, click on the "Add Task" button. A modal will appear where you can enter the task details. Once you have entered the task details, click on the "Save" button to add the task to the list.

Editing a Task

To edit a task, click on the "Edit" button next to the task you want to edit. A modal will appear where you can update the task details. Once you have made your changes, click on the "Save" button to update the task.

Deleting a Task

To delete a task, click on the "Delete" button next to the task you want to delete. The task will be removed from the list.